pyproject.toml
[project]
name = "deepimfam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
